+++ src/main/java/Jobs/MovieSuggestion.java
@@ -20,7 +20,10 @@ import java.util.Map;
 public class MovieSuggestion {
     public static void main(String[] args) throws IOException, ClassNotFoundException, InterruptedException {
         Map<String, String> options = Helper.getInputData(args);
-        Helper.deleteFolder(options.get("output"));
+
+        if (options.containsKey("overwrite") && Boolean.parseBoolean(options.get("overwrite"))) {
+            Helper.deleteFolder(options.get("output"));
+        }
 
         Job job = Job.getInstance(new Configuration(), "movieSuggestionOutput");
         job.setJarByClass(MovieSuggestion.class);
